Refactor
Refactor #5- Move class PdfDetailView, def get_best_pdf to detail.py
