+++ judgments/views/__init__.py
@@ -1,72 +1,25 @@
-import logging
-import os
 import re
 import urllib
 
-import requests
 from caselawclient.Client import (
     RESULTS_PER_PAGE,
     MarklogicAPIError,
     MarklogicResourceNotFoundError,
-    api_client,
 )
-from django.conf import settings
-from django.http import Http404, HttpResponse
-from django.shortcuts import redirect
+from django.http import Http404
 from django.template import loader
 from django.template.response import TemplateResponse
-from django.urls import reverse
 from django.utils.translation import gettext
-from django.views.generic import TemplateView
-from django_weasyprint import WeasyTemplateResponseMixin
 from ds_caselaw_utils import courts as all_courts
-from requests_toolbelt.multipart import decoder
 
 from judgments.models import SearchResult
 from judgments.utils import (
     MAX_RESULTS_PER_PAGE,
     as_integer,
-    get_pdf_uri,
     has_filters,
     paginator,
     perform_advanced_search,
 )
-
-
-class PdfDetailView(WeasyTemplateResponseMixin, TemplateView):
-    template_name = "pdf/judgment.html"
-    pdf_stylesheets = [os.path.join(settings.STATIC_ROOT, "css", "judgmentpdf.css")]
-    pdf_attachment = True
-
-    def dispatch(self, request, *args, **kwargs):
-        self.pdf_filename = f'{kwargs["judgment_uri"]}.pdf'
-
-        return super(PdfDetailView, self).dispatch(request, *args, **kwargs)
-
-    def get_context_data(self, judgment_uri, **kwargs):
-        context = super().get_context_data(**kwargs)
-
-        results = api_client.eval_xslt(judgment_uri)
-        multipart_data = decoder.MultipartDecoder.from_response(results)
-        context["judgment"] = multipart_data.parts[0].text
-
-        return context
-
-
-def get_best_pdf(request, judgment_uri):
-    """If there's a DOCX-derived PDF in the S3 bucket, return that.
-    Otherwise fall back and redirect to the weasyprint version."""
-    pdf_uri = get_pdf_uri(judgment_uri)
-    response = requests.get(pdf_uri)
-    if response.status_code == 200:
-        return HttpResponse(response.content, content_type="application/pdf")
-
-    if response.status_code != 404:
-        logging.warn(
-            f"Unexpected {response.status_code} error on {judgment_uri} whilst trying to get_best_pdf"
-        )
-    # fall back to weasy_pdf
-    return redirect(reverse("weasy_pdf", kwargs={"judgment_uri": judgment_uri}))
 
 
 def index(request):
